requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-experimental==0.0.47
pinecone-client==3.0.0
cohere==4.37
openai==1.10.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
youtube-transcript-api==0.6.2
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
rank-bm25==0.2.2
numpy==1.26.3
tiktoken==0.5.2
aiofiles==23.2.1
